+++ src/parser/features/features.js
@@ -76,6 +76,13 @@ function parseFeature(feat, ddb, character, source, type) {
     });
     item.data.description.value += `</ul>`;
     features.push(item);
+
+    const hasBackgroundType = utils.versionCompare(game.data.system.data.version, "1.6.0") >= 0;
+    if (hasBackgroundType) {
+      item.type = "background";
+      item.img = "icons/skills/trades/academics-book-study-purple.webp";
+      item.name = item.name.split("Background: ").pop();
+    }
   } else if (choices.length > 0) {
     logger.debug(`Found ${choices.map((c) => c.label).join(",")}`);
     choices.forEach((choice) => {
@@ -349,6 +356,7 @@ export default async function parseFeatures(ddb, character) {
   const backgroundSource = utils.parseSource(backgroundFeature.definition);
   const backgroundFeat = parseFeature(backgroundFeature, ddb, character, backgroundSource, "background");
   backgroundFeat.forEach((item) => {
+    console.warn("bakgroun",item)
     items.push(item);
   });